Use `window_from_bounds` fix
